+++ Runtime/KinApiCodegen/KinApiClient/src/main/CsharpDotNet2/IO/Swagger/Model/AirdropStatsDate.cs
@@ -24,7 +24,7 @@ namespace IO.Swagger.Model {
     /// </summary>
     [DataMember(Name="count", EmitDefaultValue=false)]
     [JsonProperty(PropertyName = "count")]
-    public decimal? Count { get; set; }
+    public int? Count { get; set; }
 
 
     /// <summary>
