+++ Runtime/KinApiCodegen/KinApiClient/src/main/CsharpDotNet2/IO/Swagger/Model/RequestAirdropRequest.cs
@@ -45,7 +45,7 @@ namespace IO.Swagger.Model {
     /// </summary>
     [DataMember(Name="index", EmitDefaultValue=false)]
     [JsonProperty(PropertyName = "index")]
-    public decimal? Index { get; set; }
+    public int? Index { get; set; }
 
     /// <summary>
     /// Gets or Sets Mint
